--sizes of c datatypes-hk-

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int b;
 long int c;
 long long int d;
 float e;
-printf("size of a char: %d byte(s)\n", (unsigned int)sizeof(a));
-printf("size of an int: %d byte(s)\n", (unsigned int)sizeof(b));
-printf("size of a long int: %d byte(s)\n", (unsigned int)sizeof(c));
-printf("size of a long long int: %d byte(s)\n", (unsigned int)sizeof(d));
-printf("size of a float: %d byte(s)\n", (unsigned int)sizeof(e));
+printf("size of a char: %u byte(s)\n", (unsigned long)sizeof(a));
+printf("size of an int: %u byte(s)\n", (unsigned long)sizeof(b));
+printf("size of a long int: %u byte(s)\n", (unsigned long)sizeof(c));
+printf("size of a long long int: %u byte(s)\n", (unsigned long)sizeof(d));
+printf("size of a float: %u byte(s)\n", (unsigned long)sizeof(e));
 return (0);
 }
